move types out of datafetcher
